+++ packages/evm/contracts/utils/HeaderStorage.sol
@@ -28,7 +28,7 @@ contract HeaderStorage {
     /// @notice Reverts if the given block header was not previously stored and is now out of range.
     function storeBlockHeaders(uint256[] memory blockNumbers) public returns (bytes32[] memory) {
         bytes32[] memory blockHeaders = new bytes32[](blockNumbers.length);
-        for (uint i = 0; i < blockNumbers.length; i++) {
+        for (uint256 i = 0; i < blockNumbers.length; i++) {
             blockHeaders[i] = storeBlockHeader(blockNumbers[i]);
         }
         return blockHeaders;
